#29 - Adding polaroids to blog wall
